+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [0.9.7] - Unreleased
 ### Changes
 - Add doc comment on MenuExt::value() which might return -1 when no choice is made.
+- Add WidgetExt::set_callback2 which exposes the widget in the closure arguments.
+- Remove unused safe_new.
 
 ## [0.9.6] - 2020-10-15
 ### Changes